requirements.txt
beautifulsoup4==4.13.4
openpyxl==3.1.5
pandas==2.3.2
pdfplumber==0.11.7
PyPDF2==3.0.1
pillow==11.2.1
requests==2.32.4
selenium==4.34.0
urllib3==2.4.0
webdriver-manager==4.0.2

#  Old list
#
#beautifulsoup4==4.13.4
#certifi==2025.6.15
#cryptography==45.0.4
#et_xmlfile==2.0.0
#numpy==2.3.2
#openpyxl==3.1.5
#pandas==2.3.2
#pdfkit==1.0.0
#pdfminer.six==20250506
#pdfplumber==0.11.7
#pillow==11.2.1
#pypdfium2==4.30.1
#python-dateutil==2.9.0.post0
#python-dotenv==1.1.1
#pytz==2025.2
#requests==2.32.4
#selenium==4.34.0
#urllib3==2.4.0
#webdriver-manager==4.0.2
# 